index.ts fix
