--- tools/reviewer/webapp/src/app/dashboard/diff/component-list.ts
@@ -1,11 +1,0 @@
-import { DiffDiscussionComponent } from './diff-discussion';
-import { DiffFilesComponent } from './diff-files';
-import { DiffHeaderComponentList } from './diff-header';
-import { DiffComponent } from './diff.component';
-
-export const DiffComponentList = [
-  DiffComponent,
-  ...DiffHeaderComponentList,
-  DiffFilesComponent,
-  DiffDiscussionComponent,
-];
