--- tools/reviewer/webapp/src/app/dashboard/diff/diff-header/component-list.ts
@@ -1,15 +1,0 @@
-import { AddUserDialogComponent } from './add-user-dialog';
-import { DiffHeaderContentComponent } from './diff-header-content';
-import { DiffHeaderTitlebarComponent } from './diff-header-titlebar';
-import { DiffHeaderComponent } from './diff-header.component';
-import { ReplyPopupComponent } from './reply-popup';
-import { UserPopupComponent } from './user-popup';
-
-export const DiffHeaderComponentList = [
-  DiffHeaderComponent,
-  DiffHeaderTitlebarComponent,
-  DiffHeaderContentComponent,
-  UserPopupComponent,
-  ReplyPopupComponent,
-  AddUserDialogComponent,
-];
