+++ tools/reviewer/webapp/src/app/dashboard/diff/diff-header/index.ts
@@ -4,4 +4,19 @@ export * from './diff-header-content';
 export * from './user-popup';
 export * from './reply-popup';
 export * from './add-user-dialog';
-export * from './component-list';
+
+// Components
+import { AddUserDialogComponent } from './add-user-dialog';
+import { DiffHeaderContentComponent } from './diff-header-content';
+import { DiffHeaderTitlebarComponent } from './diff-header-titlebar';
+import { DiffHeaderComponent } from './diff-header.component';
+import { ReplyPopupComponent } from './reply-popup';
+import { UserPopupComponent } from './user-popup';
+export const DiffHeaderComponentList = [
+  DiffHeaderComponent,
+  DiffHeaderTitlebarComponent,
+  DiffHeaderContentComponent,
+  UserPopupComponent,
+  ReplyPopupComponent,
+  AddUserDialogComponent,
+];
