+++ tools/reviewer/webapp/src/app/dashboard/diff/index.ts
@@ -2,4 +2,15 @@ export * from './diff.component';
 export * from './diff-header';
 export * from './diff-discussion';
 export * from './diff-files';
-export * from './component-list';
+
+// Components
+import { DiffDiscussionComponent } from './diff-discussion';
+import { DiffFilesComponent } from './diff-files';
+import { DiffHeaderComponentList } from './diff-header';
+import { DiffComponent } from './diff.component';
+export const DiffComponentList = [
+  DiffComponent,
+  ...DiffHeaderComponentList,
+  DiffFilesComponent,
+  DiffDiscussionComponent,
+];
