--- tools/reviewer/webapp/src/app/dashboard/file-changes/code-changes/component-list.ts
@@ -1,7 +1,0 @@
-import { CodeChangesComponent } from './code-changes.component';
-import { CommentsComponent } from './comments';
-
-export const CodeChangesComponentList = [
-  CodeChangesComponent,
-  CommentsComponent,
-];
