+++ tools/reviewer/webapp/src/app/dashboard/file-changes/code-changes/index.ts
@@ -2,4 +2,18 @@ export * from './code-changes.component';
 export * from './code-changes.interface';
 export * from './comments';
 export * from './services';
-export * from './component-list';
+
+// Components
+import { CodeChangesComponent } from './code-changes.component';
+import { CommentsComponent } from './comments';
+export const CodeChangesComponentList = [
+  CodeChangesComponent,
+  CommentsComponent,
+];
+
+// Services
+import { CommentsService, LineService } from './services';
+export const CodeChangesServiceList = [
+  CommentsService,
+  LineService,
+];
