--- tools/reviewer/webapp/src/app/dashboard/file-changes/component-list.ts
@@ -1,7 +1,0 @@
-import { CodeChangesComponentList } from './code-changes';
-import { FileChangesComponent } from './file-changes.component';
-
-export const FileChangesComponentList = [
-  FileChangesComponent,
-  ...CodeChangesComponentList,
-];
