--- tools/reviewer/webapp/src/app/dashboard/file-changes/service-list.ts
@@ -1,7 +1,0 @@
-import { CommentsService } from './code-changes';
-import { FileChangesService } from './file-changes.service';
-
-export const FileChangesServiceList = [
-  FileChangesService,
-  CommentsService,
-];
