--- tools/reviewer/webapp/src/app/layout/component-list.ts
@@ -1,7 +1,0 @@
-import { LayoutComponent } from './layout.component';
-import { ToolbarComponent } from './toolbar';
-
-export const LayoutComponentList = [
-  ToolbarComponent,
-  LayoutComponent,
-];
