+++ tools/reviewer/webapp/src/app/layout/index.ts
@@ -1,2 +1,9 @@
 export * from './layout.component';
-export * from './component-list';
+
+// Components
+import { LayoutComponent } from './layout.component';
+import { ToolbarComponent } from './toolbar';
+export const LayoutComponentList = [
+  ToolbarComponent,
+  LayoutComponent,
+];
