+++ tools/reviewer/webapp/src/app/page-not-found/index.ts
@@ -1,1 +1,1 @@
-export * from './not-found.component';
+export * from './page-not-found.component';
