--- tools/reviewer/webapp/src/app/shared/directives/directive-list.ts
@@ -1,5 +1,0 @@
-import { HtmlDirective } from './html.directive';
-
-export const DirectiveList = [
-  HtmlDirective,
-];
