--- tools/reviewer/webapp/src/app/shared/pipes/pipe-list.ts
@@ -1,5 +1,0 @@
-import { TimePipe } from './time.pipe';
-
-export const PipeList = [
-  TimePipe,
-];
