--- tools/reviewer/webapp/src/app/shared/services/service-list.ts
@@ -1,17 +1,0 @@
-import { AuthGuard } from './auth.guard';
-import { AuthService } from './auth.service';
-import { EncodingService } from './encoding.service';
-import { FirebaseService } from './firebase.service';
-import { HighlightService } from './highlight.service';
-import { LocalserverService } from './localserver.service';
-import { NotificationService } from './notification.service';
-
-export const ServiceList = [
-  AuthGuard,
-  AuthService,
-  FirebaseService,
-  HighlightService,
-  NotificationService,
-  EncodingService,
-  LocalserverService,
-];
